requirements.txt
flask 
sqlalchemy 
mysql-connector-python 
requests 
scikit-learn 
pandas
python-dotenv
numpy==1.26.2
scikit-surprise